[RLE-DSE] Implement load forwarding for covering value, i.e. need to make use of phi-node/silargument to forward a different values from different predecessor basic blocks
Also add a test case that were not handled by the old loadstoreopt, but by this commit in the new RLE.

rdar://22523856

Will add more LIT tests in subsequent commits.

There are still some improvements that can be made in RLE. will make those soon and
commit with performance measurements.