Split the callee analysis portion of the call graph out.
This is a simple analysis refactored from the call code. It uses the
vtables and witness tables to compute a cache of decl-to-possible-callee
mappings, and provides an interface to look these up for a
FullApplySite, as well as building and returning a singleton list for
function_ref callees.

It adds a new invalidation PreserveKind, 'Functions', which currently
nothing claims to preserve (but most everything actually does). Updating
the preservation claims of existing passes will be done as a separate
step.

This commit also adds a printing pass and test for this analysis.