Refactor bottom-up function ordering code from the call graph.
This adds a new utility as well as a test pass for computing the
bottom-up SCC ordering of the functions in the module.

Currently all functions are included, but we could consider in the
future including only those functions that are potentially reachable
from outside the current scope of compilation (module or file depending
on compile mode).

This would allow us to skip optimizing functions that we'll eventually
eliminate.