IRGen: Address-only enums should not use spare bits for layout
In the indirect initializeWithCopy value witness, we had a
terrible hack where the *source* of the initialization was
modified, stripping and emplacing the tag bits before and
after the copy of the payload, respectively.

Since address-only types should not have spare bits at all,
replace the hack with an assertion.