[AliasAnalysis] Don't cache the results between queries.
The cache is using SILValues as keys and such key would contain dangling pointers after the instruction corresponding to a given SILValue is erased.
Therefore, the cache needs to be invalidated between queries. It still makes sense to use the cache during queries, because sometimes queries may be recursive and pretty expensive to compute. It is safe to use  caching during the query, because no instructions are being erased while processing it.

The removal of the cache does not seem to affect compile times in a negative way.