Fix TBI test.
Now we perform the correct check: making sure that we have TBI on for
iOS7.0-Earlier and on for iOS8.0+Later.

The actual functionality that occurs is that we /always/ have TBI on when the
triple is an AArch64 target. Then the backend determines via the triple whether
it is supported or not for the specific os version.

I also removed the executable test bit requires since this test does not execute
anything, it just runs FileCheck tests.