Sema: Targeted fix for LLDB REPL issue
LLDB changes accessiblity of declarations after type checking,
which is not a good idea because it is likely to break invariants.
Indeed, the validateFixedLayoutAttribute() / hasFixedLayout()
logic was not prepared for this possibility.

This is a targeted fix to address the immediate breakage. A better
fix would be to change LLDB, and also to change Sema to store the
global -enable-resilience flag state in a bit in the serialized
module, instead of sticking it on every declaration.

Fixes <rdar://problem/23545959>.