[AliasAnalysis] Remove the AliasAnalysis Cache.
The AliasAnalysis is unsafe. When we delete instructions we don't have a good
mechanism for invalidating the cache and we don't want instructions to
explicitly invalidate AliasAnalysis after every deletion of an instruction.

Morever, the alias analysis cache _always_ misses. Removing the cache did not
change the build time of the standard library at all.