[PM] Make the printer/dumper passes agnostic to the order the PM scans functions in the module.
When we change the order in which the pass manager scans functions in the module
(say, bottom-up scan) the tests that inspect the output of the dumper/printer
functions break. To work around this problem I changed the dumper methods from
function passes to module passes. Now these printer passes use the order in
which the functions are defined in the test file and not rely on the
pass-manager to scan functions in a specific way.