[loop-arc] Prepare to remove opaque ARC pairing interfaces.
The ARC optimizer has been using opaque interface boundaries in between
the pairing analysis and the actual optimization pass. This was
necessary in order to not expose details from ARC pairing when the ARC
pairing computation was split in between SILAnalysis and SILPasses. Now
that is gone, I can remove this interface and simplify the layering.

The other reason to do this is that the non-loop and loop ARC optimizers
are going to start to differ a bit more since the loop optimizer is
going to use a LoopVisitor.