Remove call graph maintenance from the perf inliner.
This removes the call graph maintenance code from the performance
inliner since we've decided to go in a different direction and rather
than keeping a persistent call graph are instead splitting it up into a
few utilities that can provide the most generally useful information
from the call graph.

I had to add an unfortunate hack here due to the implicit lazy linking
that we have in the compiler, but intend on removing. In theory a change
like this should have resulted in no change in fucntionality, but in
fact not updating the call graph meant not doing some of that lazy
deserialization, which in turn changed what got inlined. This would have
been okay, but it seems to have exposed other bugs in the optimizer, so
I've added in an explicit call to a function that will deserialize newly
devirtualized function references.