Make test independent of function processing order
This test verifies copy-on-write optimizations based on array semantics happen
correctly.  We test that if SIL has the right form in cowarray_opts.sil these
optimizations happen.  But I believe that it is important to test that SIL has
the right form when we get to the cow hoisting pass. This test verifies this.
Unfortunately, this requires a swift test case.

rdar://23681223