Reapply "Make test independent of function processing order"
This reverts commit 29d6170140f8c577f91fd96da4bd3128a4e353e5.

I missed adding back the REQUIRES line.

Original commit message:

Make test independent of function processing order

This test verifies copy-on-write optimizations based on array semantics happen
correctly.  We test that if SIL has the right form in cowarray_opts.sil these
optimizations happen.  But I believe that it is important to test that SIL has
the right form when we get to the cow hoisting pass. This test verifies this.
Unfortunately, this requires a swift test case.

rdar://23681223