[docs] Commit a draft of the Library Evolution (Resilience) master doc.
There's still quite a bit of work to go here. In particular, it's not really
ready for review, but John and I think it's pointed in the right direction,
and with Slava and Joe already working in this area it's important that we
have something to anchor to. Hoping to get back to this soon.