Change Quoted Bullet Points to Normal Bullet Points
Some of the lists in this file were not formatted the correct way, and were interpreted as a quote that contained a list. The change allows all the bullet point lists to be interpreted the correct way, as just a bullet point list, and not as a quote containing a list.