SILGen: Don't emit no-payload case constructors at all
With the change to open-code enum construction, it is no longer
possible to form a reference to a no-payload case constructor
function (at least until we allow enum cases to witness protocol
requirements), so only payload case constructors need to be
emitted now.

Currently, we still emit all the case constructor functions
for payload cases unconditionally, even though now this is only
necessary for the ones that are partially applied. This will
be addressed in a follow-on patch.