[stdlib] Implement StaticString.UnicodeScalarView
Add a new type `StaticString.UnicodeScalarView`, accessible via a
property `.unicodeScalars` on `StaticString`. Also add initializers for
`StaticString` that take a `UnicodeScalarView` or a
`Slice<UnicodeScalarView>`.

The motivating reason for this is to make it possible to slice a
`StaticString` down to a substring that's still typed as `StaticString`.

Fixes rdar://problem/23382521