[Basic] Implement `CacheImpl::releaseValue`
Added implementation similar to the `getAndRetain` and `remove`
methods. Given that `DCache.Entries` appears to be a subclass of
(or similar to) `std::map`, there isn't a great way to find a
specified value without iterating through the map (unless `boost`
is included, or unless you're using boost's `bimap`). Therefore,
this works by iterating through the cache, finding the specified
value, and calling `valueDestroyCB` on it.