[AliasAnalysis] Reintroduce the cache to AliasAnalysis.
This commit reintroduces the cache to AliasAnalysis. This cache solves the
problem of dead pointers that stay in the cache maps after instructions are
being deleted. It's inefficient to scan the whole cache every time we want to
invalidate a single pointer (pointers can be a part of a key). We solve this
problem by mapping pointers to internal indices. With this map in place it's
very easy to invalidate real pointers (by removing them from the ptr-to-index
map), and there is no harm in keeping stale indices in the map. We limit the
size of the map to keep the memory usage down.

The hit rate of the cache on stdlib is about ~85%.