[arc] Create a new pass called ARCLoopOpts.
This is just a pass that runs multiple ARC loop visitors on a
function. It is necessary so that as we hoist retain, release pairs out
of loops, we then can eliminate those pairs when we process the loop
that the pair is now in.

Right now only ARCSequenceOpts is wired up, so I added a quick FileCheck
test for the pass by running it on one of the globalarcopt tests.

I am going to create a test that tests the integration in between the passes,
once I have multiple of them in place.