[PassManager] Process the function in a bottom-up order.
This commit changes the order in which we process the functions in the module
from module order to bottom-up order. Right now this change should not make any
noticeable difference, but soon we'll be able to make the inliner a function
pass and improve the caching of our inter-procedural analysis like side-effects
analysis.