Fix <rdar://problem/23718816> QoI: "Extra argument" error when accidentally currying a method
It is a somewhat common case where folks are accidentally referring to an
instance member with the Type as the base.  This forms a curried member,
which then produced head scratching error messages downstream.

Now that the prep work has gone in, the first part of this is now
straight-forward to fix: simply check for this case and diagnose it
with a custom error, which makes it more clear what the mistake was.

The other half of this problem (tracked by 22108559) affects cases where
the method you're calling takes a single argument.  This isn't fixed
yet, but I'm adding a testcase for it anyway.