Disable implicit optional up and down casts
This will be a bit controversial.

Aside from the obvious type-safety implications, this flag can be used
to debug the constraint checker, especially those paths that are
related to the generation of implicit conversions.  By turning these
off, you get a “photo-negative” picture of where the compiler is making
decisions to cast values.  In addition, projects that turn on this flag
can use it to get an accurate count of where they are making hidden
decisions about the hairier parts of optional usage and correctness.