[build-script] Create 'buildbot_linux' preset to reduce duplication
The Ubuntu 14.04 preset existed to skip building Foundation due to
unavailable ICU versions. The problem was corrected, but the
comment and preset duplication of 'buildbot_linux_*' in
build-presets.ini remained.

This change adds a 'buildbot_linux' preset as the primary Linux
preset and changes 'buildbot_linux_*' to use it as a mixin.

Implemented as discussed with Dmitri Gribenko on swift-dev.